--- src/custom_landmarks/__init__.py
@@ -1,11 +1,0 @@
-from .virtual_pose_landmark import VirtualPoseLandmark
-from .virtual_landmark import VirtualLandmark
-from .decorator import landmark
-from . import calculus
-
-__ALL__ =[
-    "VirtualPoseLandmark"
-    "VirtualLandmark",
-    "calculus"
-    "landmark"
-]
