+++ src/virtual_landmark/abstract_landmark.py
