+++ tests/test_abstract_landmark.py
@@ -1,7 +1,20 @@
+# Copyright 2024 cvpose
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import pytest
-import numpy as np
 from mediapipe.framework.formats.landmark_pb2 import NormalizedLandmark
-from custom_landmarks.abstract_landmark import AbstractLandmark
+from virtual_landmark.abstract_landmark import AbstractLandmark
 
 class DummyLandmark(AbstractLandmark):
     def add_virtual(self, name, point, connections=None):
