+++ tests/test_calculus.py
@@ -1,8 +1,21 @@
-import pytest
+# Copyright 2024 cvpose
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import numpy as np
 from mediapipe.framework.formats.landmark_pb2 import NormalizedLandmark
 
-from custom_landmarks.calculus import (
+from virtual_landmark.calculus import (
     middle, projection, centroid, mirror,
     weighted_average, extend, normalize,
     interpolate, bisector, rotate
