+++ tests/test_decorator.py
@@ -11,9 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import pytest
 
-from custom_landmarks.decorator import landmark  # adjust if your path differs
+import pytest
+from virtual_landmark import landmark
 
 
 def test_landmark_decorator_assigns_attributes():
