+++ tests/test_virtual_pose_landmark.py
@@ -1,12 +1,26 @@
+# Copyright 2024 cvpose
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import pytest
 from types import SimpleNamespace
 
-from custom_landmarks import VirtualPoseLandmark  # Substitua pelo caminho real, ex: custom_landmarks.virtual_pose_landmark
+from virtual_landmark import VirtualPoseLandmark 
 
 @pytest.fixture
 def custom_vpl():
     vpl = VirtualPoseLandmark()
-    # Limpa os dados e define um item de teste manualmente
+
     vpl.clear()
     vpl._reverse.clear()
     vpl["DUMMY"] = 99
